+++ noodles-csi/src/binning_index/index/reference_sequence/bin.rs
@@ -19,10 +19,10 @@ impl Bin {
     ///
     /// ```
     /// use noodles_csi::binning_index::index::reference_sequence::Bin;
-    /// assert_eq!(Bin::max_id(5), 37449);
+    /// assert_eq!(Bin::max_id(5), Some(37449));
     /// ```
-    pub const fn max_id(depth: u8) -> usize {
-        bin_limit(depth) as usize
+    pub const fn max_id(depth: u8) -> Option<usize> {
+        bin_limit(depth)
     }
 
     /// Calculates the metadata bin ID.
@@ -31,10 +31,13 @@ impl Bin {
     ///
     /// ```
     /// use noodles_csi::binning_index::index::reference_sequence::Bin;
-    /// assert_eq!(Bin::metadata_id(5), 37450);
+    /// assert_eq!(Bin::metadata_id(5), Some(37450));
     /// ```
-    pub const fn metadata_id(depth: u8) -> usize {
-        Self::max_id(depth) + 1
+    pub const fn metadata_id(depth: u8) -> Option<usize> {
+        match Self::max_id(depth) {
+            Some(n) => Some(n + 1),
+            None => None,
+        }
     }
 
     /// Creates a binning index reference sequence bin.
@@ -77,9 +80,11 @@ impl Bin {
 }
 
 // `CSIv1.pdf` (2020-07-21)
-const fn bin_limit(depth: u8) -> i32 {
-    assert!(depth <= 10);
-    (1 << ((depth + 1) * 3)) / 7
+const fn bin_limit(depth: u8) -> Option<usize> {
+    match 1usize.checked_shl(3 * ((depth as u32) + 1)) {
+        Some(n) => Some(n / 7),
+        None => None,
+    }
 }
 
 #[cfg(test)]